+++ src/validation.h
@@ -558,7 +558,7 @@ extern tp::ThreadPool* threadpool;
 extern std::vector<std::pair<uint256, int64_t> > vecTPSTestReceivedTimesMempool;
 extern int64_t nTPSTestingStartTime;
 extern int64_t nTPSTestingSendRawElapsedTime;
-extern int64_t nTPSTestingSendRawStartTime;
+extern double nTPSTestingSendRawStartTime;
 /**
  * Return the spend height, which is one more than the inputs.GetBestBlock().
  * While checking, GetBestBlock() refers to the parent block. (protected by cs_main)
